Event refactor

See merge request cse1105/2023-2024/teams/oopp-team-31!61

+++ client/src/main/java/client/scenes/SplittyOverviewCtrl.java
@@ -85,6 +85,8 @@ public class SplittyOverviewCtrl implements Initializable {
     @FXML
     public Label joinedEventLabel;
 
+//    private Config config;
+
     @FXML
     private ListView<Participant> participantListView;
     @Inject
@@ -113,8 +115,6 @@ public class SplittyOverviewCtrl implements Initializable {
                 }
             }
         });
-//        fetchExpenses();
-//        fetchParticipants();
     }
 
     public void setEventCode(int eventCode){
@@ -230,8 +230,9 @@ public class SplittyOverviewCtrl implements Initializable {
         expenseList.getItems().addAll(expenses);
         allExpenses.setContent(expenseList);
     }
-    public void fetchParticipants(){
 
+    public void fetchParticipants(){
+        participantListView.getItems().clear();
         List<Participant> participants = new ArrayList<>();
         try{
             participants = serverUtils.getParticipants(eventCode);
@@ -286,6 +287,17 @@ public class SplittyOverviewCtrl implements Initializable {
     }
 
     public void leaveEvent(ActionEvent actionEvent) {
+        // can only leave if balance is 0
+        Participant me = serverUtils.getParticipant(config.getId(), eventCode);
+        if (me.getBalance() != 0) {
+            Alert alert = new Alert(Alert.AlertType.ERROR);
+            alert.setTitle("Error");
+            alert.setHeaderText("You can't leave the event");
+            alert.setContentText("You can only leave the event if your balance is 0");
+            alert.showAndWait();
+            return;
+        }
+
         serverUtils.deleteParticipant(config.getId(), eventCode);
         mainCtrl.showStartScreen();
         confirmationLabel.setVisible(false);
@@ -339,6 +351,10 @@ public class SplittyOverviewCtrl implements Initializable {
         confirmationLabel.setVisible(false);
         cancelLeaveButton.setVisible(false);
         leaveConfirmationButton.setVisible(false);
+    }
+
+    public int getCurrentEventCode(){
+        return this.eventCode;
     }
 }
